Correction to incoming requests and parameter handling
The JSONRPC 2.0 specification doesn't allow for both unnamed and named parameters to be passed in simultaneously for any request. In addition, any requests that did not provide a named variable named 'kwargs' would cause an exception to be thrown. Added logic to call to the mapped method in a way that differentiates between no, unnamed or named parameters.